simplify template existence check

+++ src/import.rs
@@ -23,15 +23,6 @@ pub fn import(name: Option<String>, path: Option<String>, git_path: Option<Strin
     write_template_to_fs(template, template_folder);
 }
 
-/// Check if template exists in filesystem
-fn check_template_existence(template_path: &Path) {
-    let template_existence = Path::exists(template_path);
-
-    if template_existence {
-        panic!("Template already exists")
-    }
-}
-
 /// Write template to filesystem
 fn write_template_to_fs(template: Template, template_folder: String) {
     // Create file contents
@@ -42,7 +33,9 @@ fn write_template_to_fs(template: Template, template_folder: String) {
     let template_path = Path::new(&template_path_string);
 
     // Check if template already exists
-    check_template_existence(template_path);
+    if !template_path.exists() {
+        panic!("Template already exists");
+    }
 
     // Write template to fs ~/.config/dotfile-manager/templates/foo.toml
     let result = fs::write(template_path, toml);
